requirements.txt
# Web scraping and browser automation
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0

# Database
sqlalchemy==2.0.25

# Telegram notifications
python-telegram-bot==20.8

# Scheduling
schedule==1.2.1

# Environment variables
python-dotenv==1.0.1

# Utilities
pytz==2024.1
